Added explicit conversions to float in the point distance functions

+++ src/math/point.h
@@ -91,12 +91,12 @@ namespace Pine
 
     template <typename T>
     float distance(const Point2D<T>& p1, const Point2D<T>& p2) {
-        return std::sqrt(square(p2.x - p1.x) + square(p2.y - p1.y));
+        return std::sqrt(static_cast<float>(square(p2.x - p1.x) + square(p2.y - p1.y)));
     }
 
     template <typename T>
     float distance(const Point3D<T>& p1, const Point3D<T>& p2) {
-        return std::sqrt(square(p2.x - p1.x) + square(p2.y - p1.y) + square(p2.z - p1.z));
+        return std::sqrt(static_cast<float>(square(p2.x - p1.x) + square(p2.y - p1.y) + square(p2.z - p1.z)));
     }
 }
